Decrease compute type

+++ predict.py
@@ -9,7 +9,7 @@ os.environ['HF_HOME'] = '/src/hf_models'
 os.environ['TORCH_HOME'] = '/src/torch_models'
 
 
-compute_type = "float32"
+compute_type = "float16"
 
 
 class Predictor(BasePredictor):
